Restructure internal packages
