+++ backend/internal/oauth/oauth2/authz/codepersistence.go
@@ -24,8 +24,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/asgardeo/thunder/internal/identity/oauth2/authz/constants"
-	"github.com/asgardeo/thunder/internal/identity/oauth2/authz/model"
+	"github.com/asgardeo/thunder/internal/oauth/oauth2/authz/constants"
+	"github.com/asgardeo/thunder/internal/oauth/oauth2/authz/model"
 	"github.com/asgardeo/thunder/internal/system/database/provider"
 	"github.com/asgardeo/thunder/internal/system/log"
 )
