+++ backend/internal/oauth/oauth2/utils/oauthutils.go
@@ -25,10 +25,10 @@ import (
 	"net/url"
 	"strings"
 
-	authzmodel "github.com/asgardeo/thunder/internal/identity/oauth2/authz/model"
-	"github.com/asgardeo/thunder/internal/identity/oauth2/constants"
-	sessionmodel "github.com/asgardeo/thunder/internal/identity/session/model"
-	sessionstore "github.com/asgardeo/thunder/internal/identity/session/store"
+	authzmodel "github.com/asgardeo/thunder/internal/oauth/oauth2/authz/model"
+	"github.com/asgardeo/thunder/internal/oauth/oauth2/constants"
+	sessionmodel "github.com/asgardeo/thunder/internal/oauth/session/model"
+	sessionstore "github.com/asgardeo/thunder/internal/oauth/session/store"
 	"github.com/asgardeo/thunder/internal/system/config"
 	"github.com/asgardeo/thunder/internal/system/log"
 	"github.com/google/uuid"
