+++ backend/internal/system/utils/stringutil.go
